display text instead of value in aggregations
